ml_service/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart

# PDF parsing
pdfplumber>=0.11.0
PyMuPDF>=1.24.0

# NLP
spacy>=3.7.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl

# Embeddings
sentence-transformers>=3.0.0
torch==2.5.1+cpu ; platform_system=="Linux" and platform_machine=="x86_64"

# Vector store
faiss-cpu>=1.8.0
numpy>=1.26.0
